Tests for IDeserializedCommand.ApplyToState (#84)
* add tests for ApplyToState

* fix stryker findings

* fix stryker ignore

* update readme

* fix stryker issues

* fix todo list in stryker

now every TODO entry has a GitHub issue associated

+++ AtemSharp/State/Audio/Fairlight/FairlightAudioState.cs
@@ -14,7 +14,7 @@ public class FairlightAudioState : AudioState
     /// <remarks>
     /// Fairlight input IDs are not array indices but have semantics
     /// </remarks>
-    public Dictionary<int, FairlightAudioInput> Inputs { get; } = [];
+    public Dictionary<ushort, FairlightAudioInput> Inputs { get; } = [];
 
     public MasterProperties Master { get; } = new();
     public MonitorProperties Monitor { get; } = new();
